Added waitfor lib

+++ runner/pool.go
@@ -1,22 +1,22 @@
 package runner
 
 type Pool struct {
-	size   int
+	size   uint64
 	values chan struct{}
 }
 
-func NewPool(size int) *Pool {
+func NewPool(size uint64) *Pool {
 	return &Pool{
 		size:   size,
 		values: make(chan struct{}, size),
 	}
 }
 
-func (pool *Pool) Capacity() int {
-	return pool.size - len(pool.values)
+func (pool *Pool) Capacity() uint64 {
+	return pool.size - uint64(len(pool.values))
 }
 
-func (pool *Pool) Size() int {
+func (pool *Pool) Size() uint64 {
 	return pool.size
 }
